feat: Port Quasar Shader Engine to Minecraft 1.21.11 (Fabric)
- Target Minecraft 1.21.11 using Java 21.
- Preserved Quasar branding (quasar mod ID, Quasar display name, quasar.properties).
- Adapted QuasarContext for GLES mobile GPU driver detection (ltw, Mali, Adreno, etc.).
- Integrated QuasarRecoveryLadder in GlShader compilation.
- Added float texture format downgrades and MRT draw buffer clamping.
- Added custom Quasar uniforms (quasar_RenderScale, quasar_ESVersion, quasar_DeviceTier, quasar_MV, quasar_P, quasar_MVP).
- Re-targeted Quasar GUI screens to 1.21.11 Mojang-mapped GuiGraphics APIs.

+++ common/src/main/java/net/irisshaders/iris/shaderpack/properties/PackRenderTargetDirectives.java
@@ -57,7 +57,7 @@ public class PackRenderTargetDirectives {
 
 		renderTargetSettings.forEach((index, settings) -> {
 			if (settings.shouldClear()) {
-				buffersToBeCleared.add(index);
+				buffersToBeCleared.add(index.intValue());
 			}
 		});
